Array sorting and debugging
I was able to add all the buttons to change the array in the guest book.It can delete the first and the last last. It also sorts the array from A to Z and Z to A. I started to add the ability to delete individual spots but it is still buggy.